--- src/operandi_server/operandi_server/cli/__init__.py
@@ -1,16 +1,0 @@
-import click
-
-__all__ = ['cli']
-
-from .server import server_cli
-
-
-@click.group()
-@click.version_option()
-def cli(**kwargs):  # pylint: disable=unused-argument
-    """
-    Entry-point of multipurpose CLI for Operandi Server
-    """
-
-
-cli.add_command(server_cli)
