--- src/operandi_server/operandi_server/models.py
@@ -1,2 +1,0 @@
-# Specific response models of the OPERANDI server will be implemented here
-# TODO: Implement proper response models - when integrating OCR-D WebAPI
