--- src/service_broker/service_broker/cli/__init__.py
@@ -1,16 +1,0 @@
-import click
-
-__all__ = ['cli']
-
-from .broker import broker_cli
-
-
-@click.group()
-@click.version_option()
-def cli(**kwargs):  # pylint: disable=unused-argument
-    """
-    Entry-point of multipurpose CLI for Broker
-    """
-
-
-cli.add_command(broker_cli)
